+++ package.json
@@ -2,7 +2,7 @@
   "name": "cessda.cdc.fuji.runner",
   "version": "1.0.0",
   "description": "API Runner for CDC FAIR assessment",
-  "main": "src/assessSitemaps.ts",
+  "main": "src/sitemaps.ts",
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
