+++ src/helpers/studiesAssessment.ts
@@ -8,19 +8,6 @@ import { resultsToElastic } from './esFunctions.js';
 import { existsSync, mkdirSync, writeFileSync } from 'fs';
 import { logger } from "./logger.js";
 
-//To use if testing against CDC staging
-const cdcusername = process.env['CDC_USERNAME'];
-const cdcpassword = process.env['CDC_PASSWORD'];
-
-const usernamePasswordBuffer = Buffer.from(
-    `${cdcusername}:${cdcpassword}`,
-    "utf-8"
-);
-const base64UsernamePassword = usernamePasswordBuffer.toString("base64");
-const requestHeaders = {
-    Authorization: `Basic ${base64UsernamePassword}`,
-};
-
 export async function getStudiesAssess(studiesAssessFiltered: string[], hostname: string): Promise<void> {
     //create date of testing
     const runDate = new Date();
@@ -51,7 +38,7 @@ export async function getStudiesAssess(studiesAssessFiltered: string[], hostname
             //get the study info from CDC Internal API
             studyInfo.fileName = studyInfo.urlParams?.get('q') + "-" + studyInfo.urlParams?.get('lang') + "-" + studyInfo.assessDate;
             studyInfo.cdcID = studyInfo.urlParams?.get('q');
-            const temp: StudyInfo = await getCDCApiInfo(studyInfo, requestHeaders);
+            const temp: StudyInfo = await getCDCApiInfo(studyInfo);
             studyInfo.publisher = temp.publisher;
             studyInfo.cdcStudyNumber = temp.cdcStudyNumber;
             studyInfo.oaiLink = site.includes("datacatalogue.cessda.eu") ? "https://datacatalogue.cessda.eu/oai-pmh/v0/oai" : "https://datacatalogue-staging.cessda.eu/oai-pmh/v0/oai";
@@ -92,7 +79,7 @@ export async function getStudiesAssess(studiesAssessFiltered: string[], hostname
             }
         }
         //get results from EVA and FUJI API
-        let [evaData, fujiData] = await Promise.allSettled([getEVAResults(studyInfo), getFUJIResults(studyInfo, base64UsernamePassword)]);
+        let [evaData, fujiData] = await Promise.allSettled([getEVAResults(studyInfo), getFUJIResults(studyInfo)]);
         //const fujiData1: JSON | string = await getFUJIResults(studyInfo, base64UsernamePassword); //get results from FUJI API
         resultsToElastic(studyInfo.fileName + "-EVA", evaData);
         resultsToHDD(dir, studyInfo.fileName + "-EVA.json", evaData);
